+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { compare, genSalt, hash } from 'bcrypt';
@@ -10,7 +10,8 @@ import { User } from './entities/user.entity';
 export class UsersService {
   constructor(@InjectRepository(User) private usersRepository: Repository<User>) {}
 
-  getAllUsers(): Promise<User[]> {
+  async getAllUsers(requestingUserId: number): Promise<User[]> {
+    await this.checkAdmin(requestingUserId);
     return this.usersRepository.find();
   }
 
@@ -26,7 +27,9 @@ export class UsersService {
     return await this.usersRepository.save(user);
   }
 
-  async getUserById(id: number): Promise<User> {
+  async getUserById(id: number, requestingUserId: number): Promise<User> {
+    await this.checkAdminOrOwnUser(id, requestingUserId);
+
     const user = await this.usersRepository.findOne(id);
     if (!user) throw new NotFoundException('User not found for given id');
     return user;
@@ -37,7 +40,9 @@ export class UsersService {
     return user;
   }
 
-  async updateUserById(id: number, attrs: Partial<User>): Promise<User> {
+  async updateUserById(id: number, requestingUserId: number, attrs: Partial<User>): Promise<User> {
+    await this.checkAdminOrOwnUser(id, requestingUserId);
+
     const user = await this.usersRepository.findOne(id);
     if (!user) throw new NotFoundException('User not found for given id');
 
@@ -50,7 +55,9 @@ export class UsersService {
     return this.usersRepository.save(user);
   }
 
-  async removeUserById(id: number): Promise<User> {
+  async removeUserById(id: number, requestingUserId: number): Promise<User> {
+    await this.checkAdminOrOwnUser(id, requestingUserId);
+
     const user = await this.usersRepository.findOne(id);
     if (!user) throw new NotFoundException('User not found for given id');
     return await this.usersRepository.remove(user);
@@ -60,9 +67,24 @@ export class UsersService {
     const isMatch = await compare(password, hashedPassword);
     return isMatch;
   }
+
   async hashPassword(password: string): Promise<string> {
     const salt = await genSalt();
     const hashedPassword = await hash(String(password), salt);
     return hashedPassword;
+  }
+
+  async checkAdmin(requestingUserId: number): Promise<boolean> {
+    const { isAdmin } = await this.usersRepository.findOne(requestingUserId, { select: ['isAdmin'] });
+    if (!isAdmin) throw new ForbiddenException('You have no right of access!');
+    return isAdmin;
+  }
+
+  async checkAdminOrOwnUser(id: number, requestingUserId: number): Promise<boolean> {
+    const isAdmin = await this.checkAdmin(requestingUserId);
+    const isOwnUser = id === requestingUserId;
+    const hasRights = isAdmin || isOwnUser;
+    if (!hasRights) throw new ForbiddenException('You have no right of access!');
+    return hasRights;
   }
 }
